# 13 fix column type "metric.value" to double precision as recommended

+++ migrations/01_metric.sql
@@ -3,5 +3,5 @@ CREATE TABLE IF NOT EXISTS metric
     id   VARCHAR(50) PRIMARY KEY,
     type VARCHAR(50),
     delta bigint,
-    value decimal
+    value double precision
 );
